interconnecting with the hisysevent
Match-id-a4f5bd1bafb010a0650b76aaa26f48a34150d08a

+++ frameworks/common/constants/include/security_guard_define.h
@@ -30,7 +30,7 @@ typedef struct EventInfoSt {
     int64_t eventId;
     const char *version;
     uint32_t contentLen;
-    uint32_t content[CONTENT_MAX_LEN];
+    uint8_t content[CONTENT_MAX_LEN];
 } EventInfoSt;
 
 typedef struct DeviceIdentify {
